Pull out members to utils and constants files
