+++ constants.js
@@ -1,3 +1,4 @@
 export const OPTIONS_DOMAIN_EXCLUSIONS = "domain_exclusions";
 export const STATE_OPEN_DOMAINS = "open_domains";
-
+export const WIKI_URL="https://wiki.rossmanngroup.com/wiki";
+export const PAGES_DB_JSON_URL = "https://raw.githubusercontent.com/WayneKeenan/ClintonCAT/refs/heads/main/pages_db.json";
